Code seperation done as per the article and code readability
