+++ csharp-advanced-topics/ActionAndFuncDelegatesInCsharp/ActionAndFuncDelegatesInCsharp.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.6.33723.286
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionDelegates", "ActionDelegates\ActionDelegates.csproj", "{D8F0ED26-BDB5-4029-9F6D-C5455AB929E7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActionAndFuncDelegatesUnitTests", "ActionAndFuncDelegatesUnitTests\ActionAndFuncDelegatesUnitTests.csproj", "{C5E79787-A503-4B83-B934-348787597939}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FuncDelegates", "FuncDelegates\FuncDelegates.csproj", "{76EB57AD-51BD-4AC3-8825-267B46A2F68D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D8F0ED26-BDB5-4029-9F6D-C5455AB929E7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D8F0ED26-BDB5-4029-9F6D-C5455AB929E7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D8F0ED26-BDB5-4029-9F6D-C5455AB929E7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D8F0ED26-BDB5-4029-9F6D-C5455AB929E7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5E79787-A503-4B83-B934-348787597939}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5E79787-A503-4B83-B934-348787597939}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5E79787-A503-4B83-B934-348787597939}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5E79787-A503-4B83-B934-348787597939}.Release|Any CPU.Build.0 = Release|Any CPU
+		{76EB57AD-51BD-4AC3-8825-267B46A2F68D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{76EB57AD-51BD-4AC3-8825-267B46A2F68D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{76EB57AD-51BD-4AC3-8825-267B46A2F68D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{76EB57AD-51BD-4AC3-8825-267B46A2F68D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4F6A650E-1A25-4CF4-B238-7C3F825546ED}
+	EndGlobalSection
+EndGlobal
